Refactoring: next page 타입 수정
