+++ pages/posts/index.tsx
@@ -1,14 +1,14 @@
 import axios from "axios";
-import { GetStaticProps, InferGetStaticPropsType } from "next";
+import { GetStaticProps, InferGetStaticPropsType, NextPage } from "next";
 import Posts from "../../components/posts/Posts";
 import { IPost } from "../../utils/types";
 import Text from "../../components/css_components/Text";
 import React from "react";
 import router from "next/router";
 
-const PostsPage = ({
+const PostsPage: NextPage<InferGetStaticPropsType<typeof getStaticProps>> = ({
   posts,
-}: InferGetStaticPropsType<typeof getStaticProps>) => {
+}) => {
   return (
     <>
       <Posts posts={posts} />
@@ -20,8 +20,7 @@ export default PostsPage;
 
 const getPosts = async () => {
   const res = await axios.get("http://localhost:3000/api/posts/getPosts"); // baseURL 까지 명시해야 함
-  const data = await res.data;
-  return data;
+  return res.data;
 };
 
 export const getStaticProps: GetStaticProps = async () => {
