+++ pages/write/index.tsx
@@ -1,6 +1,7 @@
+import { NextPage } from "next";
 import Write from "../../components/write/Write";
 
-const WritePage = () => {
+const WritePage: NextPage = () => {
   return (
     <>
       <Write />
